// ui/src/openapi/models/ApiProblemForEdit.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ShpanKids API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApiProblemAnswerForEdit } from './ApiProblemAnswerForEdit';
import {
    ApiProblemAnswerForEditFromJSON,
    ApiProblemAnswerForEditFromJSONTyped,
    ApiProblemAnswerForEditToJSON,
} from './ApiProblemAnswerForEdit';

/**
 * 
 * @export
 * @interface ApiProblemForEdit
 */
export interface ApiProblemForEdit {
    /**
     * 
     * @type {string}
     * @memberof ApiProblemForEdit
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProblemForEdit
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProblemForEdit
     */
    description?: string;
    /**
     * 
     * @type {Array<ApiProblemAnswerForEdit>}
     * @memberof ApiProblemForEdit
     */
    answers: Array<ApiProblemAnswerForEdit>;
}

/**
 * Check if a given object implements the ApiProblemForEdit interface.
 */
export function instanceOfApiProblemForEdit(value: object): boolean {
    if (!('title' in value)) return false;
    if (!('answers' in value)) return false;
    return true;
}

export function ApiProblemForEditFromJSON(json: any): ApiProblemForEdit {
    return ApiProblemForEditFromJSONTyped(json, false);
}

export function ApiProblemForEditFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiProblemForEdit {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'answers': ((json['answers'] as Array<any>).map(ApiProblemAnswerForEditFromJSON)),
    };
}

export function ApiProblemForEditToJSON(value?: ApiProblemForEdit | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'answers': ((value['answers'] as Array<any>).map(ApiProblemAnswerForEditToJSON)),
    };
}